requirements.txt
absl-py==0.8.0
astor==0.8.0
astroid==2.2.5
backports.weakref==1.0rc1
bleach==1.5.0
boto3==1.9.197
botocore==1.12.197
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
dlib==19.17.0
docutils==0.14
face-recognition==1.2.3
face-recognition-models==0.3.0
ffmpeg==1.4
Flask==1.0.3
gast==0.2.2
get==2019.4.13
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
html5lib==0.9999999
idna==2.8
imutils==0.5.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonpath==0.82
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.17.1
opencv-python==4.1.0.25
pdf2image==1.6.0
Pillow==6.0.0
post==2019.4.13
protobuf==3.9.1
public==2019.4.13
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.0
six==1.12.0
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.25.3
uWSGI==2.0.18
Werkzeug==0.15.5
wrapt==1.11.2
